+++ src/Settings.ts
@@ -9,12 +9,20 @@ class Settings {
         workspace.getConfiguration('dmbinder').update('homebrewPreviewEnabled', val);
     }
 
+    public async updateHomebreweryEnabled(val: boolean | undefined): Promise<void> {
+        return await workspace.getConfiguration('dmbinder').update('homebrewPreviewEnabled', val);
+    }
+
     public get treeViewStyle(): string | undefined {
         return workspace.getConfiguration('dmbinder').get('treeViewStyle');
     }
 
     public set treeViewStyle(val: string | undefined) {
         workspace.getConfiguration('dmbinder').update('treeViewStyle', val);
+    }
+
+    public async updateTreeViewStyle(val: string | undefined): Promise<void> {
+        return await workspace.getConfiguration('dmbinder').update('treeViewStyle', val);
     }
 
     public get chromeExecutablePath(): string | undefined {
@@ -25,8 +33,16 @@ class Settings {
         workspace.getConfiguration('dmbinder').update('chromeExecutablePath', val);
     }
 
+    public async updateChromeExecutablePath(val: string | undefined): Promise<void> {
+        return await workspace.getConfiguration('dmbinder').update('chromeExecutablePath', val);
+    }
+
     public get generateGettingStartedEnabled(): boolean | undefined {
         return workspace.getConfiguration('dmbinder').get('generateGettingStartedEnabled');
+    }
+
+    public async updateGenerateGettingStartedEnabled(val: boolean | undefined): Promise<void> {
+        return await workspace.getConfiguration('dmbinder').update('generateGettingStartedEnabled', val);
     }
 }